Shipping fixed
Shipping now shows up on the CheckOut page.
When the user clicks on Check-Out, the following will occur:
- a call to product.ship_prices_for(@current_user)
which will loop thru the number of items the user
has in his/her cart and calculate shipping cost for
each item.
- product.ship_prices_for(@current_user) then
returns 1 total price for each of the shipping
options based on the number of items being
purchased.

This will then be displayed in the Make Purchase page as Shipping Cost +
Sub-Total.

I'm still working thru dressing up the GUI but this is what it looks
like right now.